Refactor app.ts into "/api" and "/lib"
